Improve API of `opt` module
