+++ lib/src/api/opt/mod.rs
@@ -5,6 +5,8 @@ pub mod auth;
 mod query;
 mod resource;
 mod server_addrs;
+mod strict;
+mod tls;
 
 use crate::api::err::Error;
 use crate::api::Result;
@@ -20,7 +22,10 @@ use std::collections::BTreeMap;
 
 pub use query::*;
 pub use resource::*;
-pub use server_addrs::*;
+pub use strict::*;
+pub use tls::*;
+
+pub(crate) use server_addrs::*;
 
 /// Record ID
 pub type RecordId = Thing;
