+++ lib/src/api/opt/server_addrs/mod.rs
@@ -18,32 +18,6 @@ use crate::api::Connection;
 use crate::api::Result;
 use url::Url;
 
-/// TLS Configuration
-#[cfg(any(feature = "native-tls", feature = "rustls"))]
-#[cfg_attr(docsrs, doc(cfg(any(feature = "native-tls", feature = "rustls"))))]
-#[derive(Debug)]
-pub enum Tls {
-	/// Native TLS configuration
-	#[cfg(feature = "native-tls")]
-	#[cfg_attr(docsrs, doc(cfg(feature = "native-tls")))]
-	Native(native_tls::TlsConnector),
-	/// Rustls configuration
-	#[cfg(feature = "rustls")]
-	#[cfg_attr(docsrs, doc(cfg(feature = "rustls")))]
-	Rust(rustls::ClientConfig),
-}
-
-/// Enables `strict` server mode
-#[cfg(any(
-	feature = "kv-mem",
-	feature = "kv-tikv",
-	feature = "kv-rocksdb",
-	feature = "kv-fdb",
-	feature = "kv-indxdb",
-))]
-#[derive(Debug)]
-pub struct Strict;
-
 /// A server address used to connect to the server
 #[derive(Debug)]
 #[allow(dead_code)] // used by the embedded and remote connections
@@ -52,7 +26,7 @@ pub struct ServerAddrs {
 	#[allow(dead_code)] // used by the embedded database
 	pub(crate) strict: bool,
 	#[cfg(any(feature = "native-tls", feature = "rustls"))]
-	pub(crate) tls_config: Option<Tls>,
+	pub(crate) tls_config: Option<super::Tls>,
 }
 
 /// A trait for converting inputs to a server address object
