+++ lib/src/api/opt/strict.rs
@@ -1,0 +1,10 @@
+/// Enables `strict` server mode
+#[cfg(any(
+	feature = "kv-mem",
+	feature = "kv-tikv",
+	feature = "kv-rocksdb",
+	feature = "kv-fdb",
+	feature = "kv-indxdb",
+))]
+#[derive(Debug)]
+pub struct Strict;
